+++ src/include/mm/planemap.h
@@ -28,11 +28,130 @@ namespace mm {
     size_type y;
   };
 
+  struct fast_position {
+    typedef typename position::size_type size_type;
+    size_type index;
+  };
+
+  class position_iterator {
+  public:
+    position_iterator(const position_iterator&) = default;
+    position_iterator& operator=(const position_iterator&) = default;
+
+    fast_position operator*() {
+      return { m_index };
+    }
+
+    position_iterator& operator++() {
+      ++m_index;
+      return *this;
+    }
+
+    bool operator!=(const position_iterator& other) const {
+      return m_index != other.m_index;
+    }
+
+  private:
+    friend class position_range;
+
+    position_iterator(fast_position::size_type index)
+    : m_index(index)
+    {
+    }
+
+    fast_position::size_type m_index;
+  };
+
+  class position_range {
+  public:
+    position_range(const position_range&) = default;
+    position_range& operator=(const position_range&) = default;
+
+    position_iterator begin() {
+      return { m_b };
+    }
+
+    position_iterator end() {
+      return { m_e };
+    }
+
+  private:
+    template<class T, class Allocator>
+    friend class planemap;
+
+    position_range(fast_position::size_type b, fast_position::size_type e)
+    : m_b(b)
+    , m_e(e)
+    {
+    }
+
+    fast_position::size_type m_b;
+    fast_position::size_type m_e;
+  };
+
+
+  class index_iterator {
+  public:
+    typedef typename position::size_type size_type;
+
+    index_iterator(const index_iterator&) = default;
+    index_iterator& operator=(const index_iterator&) = default;
+
+    size_type operator*() {
+      return m_index;
+    }
+
+    index_iterator& operator++() {
+      ++m_index;
+      return *this;
+    }
+
+    bool operator!=(const index_iterator& other) const {
+      return m_index != other.m_index;
+    }
+
+  private:
+    friend class index_range;
+
+    index_iterator(size_type index)
+    : m_index(index)
+    {
+    }
+
+    size_type m_index;
+  };
+
+  class index_range {
+  public:
+    index_range(const index_range&) = default;
+    index_range& operator=(const index_range&) = default;
+
+    index_iterator begin() {
+      return { m_b };
+    }
+
+    index_iterator end() {
+      return { m_e };
+    }
+
+  private:
+    template<class T, class Allocator>
+    friend class planemap;
+
+    index_range(index_iterator::size_type b, index_iterator::size_type e)
+    : m_b(b)
+    , m_e(e)
+    {
+    }
+
+    index_iterator::size_type m_b;
+    index_iterator::size_type m_e;
+  };
+
   struct size_only_t {
   };
 
   constexpr size_only_t size_only = size_only_t();
-
 
   template<class T, class Allocator =  std::allocator<T>>
   class planemap {
@@ -167,6 +286,14 @@ namespace mm {
       return at(pos.x, pos.y);
     }
 
+    reference at(fast_position pos) {
+      if (pos.index >= m_w * m_h) {
+        throw std::out_of_range("planemap::at");
+      }
+
+      return m_content[pos.index];
+    }
+
     const_reference at(size_type x, size_type y) const {
       if (x >= m_w || y >= m_h) {
         throw std::out_of_range("planemap::at");
@@ -179,12 +306,24 @@ namespace mm {
       return at(pos.x, pos.y);
     }
 
+    const_reference at(fast_position pos) const {
+      if (pos.index >= m_w * m_h) {
+        throw std::out_of_range("planemap::at");
+      }
+
+      return m_content[pos.index];
+    }
+
     reference operator()(size_type x, size_type y) {
       return get(x, y);
     }
 
     reference operator()(position pos) {
-      return get(pos.x, pos.y);
+      return get(pos);
+    }
+
+    reference operator()(fast_position pos) {
+      return m_content[pos.index];
     }
 
     const_reference operator()(size_type x, size_type y) const {
@@ -192,7 +331,11 @@ namespace mm {
     }
 
     const_reference operator()(position pos) const {
-      return get(pos.x, pos.y);
+      return get(pos);
+    }
+
+    const_reference operator()(fast_position pos) const {
+      return m_content[pos.index];
     }
 
     // capacity
@@ -231,10 +374,175 @@ namespace mm {
       std::swap(m_content, other.m_content);
     }
 
+    // visitors
+
+    template<typename Func>
+    void visit4neighbours(size_type x, size_type y, Func func) {
+      if (x > 0) {
+        position pos{x - 1, y};
+        func(pos, get(pos));
+      }
+
+      if (x < m_w - 1) {
+        position pos{x + 1, y};
+        func(pos, get(pos));
+      }
+
+      if (y > 0) {
+        position pos{x, y - 1};
+        func(pos, get(pos));
+      }
+
+      if (y < m_h - 1) {
+        position pos{x, y + 1};
+        func(pos, get(pos));
+      }
+    }
+
+    template<typename Func>
+    void visit4neighbours(position pos, Func func) {
+      visit4neighbours(pos.x, pos.y, func);
+    }
+
+    template<typename Func>
+    void visit4neighbours(size_type x, size_type y, Func func) const {
+      if (x > 0) {
+        position pos{x - 1, y};
+        func(pos, get(pos));
+      }
+
+      if (x < m_w - 1) {
+        position pos{x + 1, y};
+        func(pos, get(pos));
+      }
+
+      if (y > 0) {
+        position pos{x, y - 1};
+        func(pos, get(pos));
+      }
+
+      if (y < m_h - 1) {
+        position pos{x, y + 1};
+        func(pos, get(pos));
+      }
+    }
+
+    template<typename Func>
+    void visit4neighbours(position pos, Func func) const {
+      visit4neighbours(pos.x, pos.y, func);
+    }
+
+
+    template<typename Func>
+    void visit8neighbours(size_type x, size_type y, Func func) {
+      for (int i = -1; i <= 1; ++i) {
+        if (x == 0 && i == -1) {
+          continue;
+        }
+
+        if (x == m_w - 1 && i == 1) {
+          continue;
+        }
+
+        for (int j = -1; j <= 1; ++j) {
+          if (y == 0 && j == -1) {
+            continue;
+          }
+
+          if (y == m_h - 1 && j == 1) {
+            continue;
+          }
+
+          if (i != 0 || j != 0) {
+            position pos{x + i, y + j};
+            func(pos, get(pos));
+          }
+        }
+      }
+    }
+
+    template<typename Func>
+    void visit8neighbours(position pos, Func func) {
+      visit8neighbours(pos.x, pos.y, func);
+    }
+
+    template<typename Func>
+    void visit8neighbours(size_type x, size_type y, Func func) const {
+      for (int i = -1; i <= 1; ++i) {
+        if (x == 0 && i == -1) {
+          continue;
+        }
+
+        if (x == m_w - 1 && i == 1) {
+          continue;
+        }
+
+        for (int j = -1; j <= 1; ++j) {
+          if (y == 0 && j == -1) {
+            continue;
+          }
+
+          if (y == m_h - 1 && j == 1) {
+            continue;
+          }
+
+          if (i != 0 || j != 0) {
+            position pos{x + i, y + j};
+            func(pos, get(pos));
+          }
+        }
+      }
+    }
+
+    template<typename Func>
+    void visit8neighbours(position pos, Func func) const {
+      visit8neighbours(pos.x, pos.y, func);
+    }
+
+    // modifiers
+
+    void reset(value_type value) {
+      size_type end = m_w * m_h;
+
+      for (size_type i = 0; i < end; ++i) {
+        m_content[i] = value;
+      }
+    }
+
+    // utils
+
+    position_range positions() const {
+      return { 0, m_w * m_h };
+    }
+
+    position to_position(fast_position pos) const {
+      return { pos.index / m_h, pos.index % m_h };
+    }
+
+    index_range x_range() const {
+      return { 0, m_w };
+    }
+
+    index_range y_range() const {
+      return { 0, m_h };
+    }
+
   protected:
 
-    reference get(size_type x, size_type y) const {
+    reference get(size_type x, size_type y) {
       return m_content[x * m_h + y];
+    }
+
+    const_reference get(size_type x, size_type y) const {
+      return m_content[x * m_h + y];
+    }
+
+    reference get(position pos) {
+      return get(pos.x, pos.y);
+    }
+
+    const_reference get(position pos) const {
+      return get(pos.x, pos.y);
     }
 
   private: