+++ src/include/mm/playability.h
@@ -13,28 +13,40 @@
  * ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
  * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  */
-#ifndef MM_GAUSSIZE_H
-#define MM_GAUSSIZE_H
+#ifndef MM_PLAYABILITY_H
+#define MM_PLAYABILITY_H
 
+#include <tuple>
+
+#include <mm/binarymap.h>
 #include <mm/heightmap.h>
+#include <mm/reachability.h>
 
 namespace mm {
 
-  class gaussize {
+  class playability {
   public:
-    typedef std::size_t size_type;
-
-    gaussize(double spread)
-    : m_spread(spread)
+    playability(double sea_level, reachability::size_type unit_size, reachability::size_type building_size, double unit_talus, double building_talus, bool output_intermediates)
+    : m_sea_level(sea_level)
+    , m_unit_size(unit_size)
+    , m_building_size(building_size)
+    , m_unit_talus(unit_talus)
+    , m_building_talus(building_talus)
+    , m_output_intermediates(output_intermediates)
     {
     }
 
-    heightmap operator()(const heightmap& src) const;
+    std::tuple<binarymap, binarymap, binarymap> operator()(const heightmap& src) const;
 
   private:
-    double m_spread;
+      double m_sea_level;
+      reachability::size_type m_unit_size;
+      reachability::size_type m_building_size;
+      double m_unit_talus;
+      double m_building_talus;
+      bool m_output_intermediates;
   };
 
 }
 
-#endif // MM_GAUSSIZE_H
+#endif // MM_PLAYABILITY_H