+++ src/include/mm/utils.h
@@ -13,45 +13,35 @@
  * ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
  * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  */
-#include <mm/accessibility.h>
-
-#include <mm/invert.h>
+#ifndef MM_UTILS_H
+#define MM_UTILS_H
 
 namespace mm {
 
-  binarymap accessibility::operator()(const binarymap& src) {
-    binarymap visited = invert()(src);
-
-    position best_pos{0, 0};
-    binarymap::size_type best_count = 0;
-
-    for (auto x : visited.x_range()) {
-      for (auto y : visited.y_range()) {
-        position pos{x, y};
-
-        if (visited(pos)) {
-          continue;
-        }
-
-        binarymap::size_type count = visited.walk(pos, nullptr);
-
-        if (count > best_count) {
-          best_pos = pos;
-          best_count = count;
-        }
-      }
+  inline
+  double value_with_sea_level(double value, double sea_level) {
+    if (value < sea_level) {
+      return value / sea_level * 0.5;
     }
 
-    visited = invert()(src);
+    return (value - sea_level) / (1.0 - sea_level) * 0.5 + 0.5;
+  }
 
-    binarymap map(src.width(), src.height(), false);
+  inline
+  double clamp(double value, double min, double max) {
+    if (value < min) {
+      return min;
+    }
 
-    map(best_pos) = true;
-    visited.walk(best_pos, [&map](position pos) {
-      map(pos) = true;
-    });
+    if (value > max) {
+      return max;
+    }
 
-    return map;
+    return value;
   }
 
 }
+
+#endif // MM_UTILS_H
+
+