+++ src/lib/color_ramp.cc
@@ -15,7 +15,12 @@
  */
 #include <mm/color_ramp.h>
 
+#include <cassert>
+#include <algorithm>
+#include <string>
+
 #include <mm/curve.h>
+#include <mm/utils.h>
 
 namespace mm {
 
@@ -45,7 +50,7 @@ namespace mm {
     m_map.emplace(offset, c);
   }
 
-  color color_ramp::get_color(double offset) const {
+  color color_ramp::compute_color(double offset) const {
     if (offset < m_min || offset > m_max) {
       return color();
     }
@@ -65,6 +70,68 @@ namespace mm {
     }
 
     return lerp(c1, c2, (offset - t1) / (t2 - t1));
+  }
+
+#if 0
+  biomeset color_ramp::compute_biomeset() const {
+    std::vector<double> limits;
+
+    double prev = -1.0;
+    for (auto value : m_map) {
+      if (prev < 0) {
+        prev = value.first;
+      } else {
+        double curr = value.first;
+        double limit = (prev + curr) / 2.0;
+        prev = curr;
+        limits.push_back(limit);
+      }
+    }
+    limits.push_back(1.0);
+    assert(m_map.size() == limits.size());
+    assert(std::is_sorted(limits.begin(), limits.end()));
+
+    biomeset set;
+    set.add_terrain({ "River", { 19, 149, 255 }}).add_climate({ 0.5, 1 }, { 0, 1 }, true); // river
+
+    prev = 0.0;
+    std::size_t index = 0;
+    for (auto value : m_map) {
+      std::string name = "Biome #" + std::to_string(index);
+
+      double curr = limits[index];
+
+      if (prev < 0.5 && 0.5 < curr) {
+        set.add_terrain({ name, value.second })
+            .add_climate({ prev, 0.5 }, { 0, 1 }, true)
+            .add_climate({ 0.5, curr }, { 0, 1 }, false);
+      } else {
+        set.add_terrain({ name, value.second })
+            .add_climate({ prev, curr }, { 0, 1 }, (curr < 0.5));
+      }
+
+      prev = curr;
+      index++;
+    }
+
+    return set;
+  }
+#endif
+
+  color_ramp color_ramp::basic() {
+    // see: http://www.blitzbasic.com/codearcs/codearcs.php?code=2415
+    color_ramp ramp;
+    ramp.add_color_stop(0.000, {  2,  43,  68}); // very dark blue: deep water
+    ramp.add_color_stop(0.250, {  9,  62,  92}); // dark blue: water
+    ramp.add_color_stop(0.490, { 17,  82, 112}); // blue: shallow water
+    ramp.add_color_stop(0.500, { 69, 108, 118}); // light blue: shore
+    ramp.add_color_stop(0.501, { 42, 102,  41}); // green: grass
+    ramp.add_color_stop(0.750, {115, 128,  77}); // light green: veld
+    ramp.add_color_stop(0.850, {153, 143,  92}); // brown: tundra
+    ramp.add_color_stop(0.950, {179, 179, 179}); // grey: rocks
+    ramp.add_color_stop(1.000, {255, 255, 255}); // white: snow
+
+    return ramp;
   }
 
 }