+++ src/lib/cutoff.cc
@@ -29,7 +29,7 @@ namespace mm {
       }
     }
 
-    return std::move(map);
+    return map;
   }
 
 }