+++ src/lib/fractal.cc
@@ -39,7 +39,7 @@ namespace mm {
       }
     }
 
-    return std::move(map);
+    return map;
   }
 
 }