+++ src/lib/hydraulic_erosion.cc
@@ -135,7 +135,7 @@ namespace mm {
 
     }
 
-    return std::move(map);
+    return map;
   }
 
 }