+++ src/lib/reachability.cc
@@ -40,7 +40,7 @@ namespace mm {
       }
     }
 
-    return std::move(map);
+    return map;
   }
 
 }