+++ src/lib/thermal_erosion.cc
@@ -72,7 +72,7 @@ namespace mm {
       }
     }
 
-    return std::move(map);
+    return map;
   }
 
 }