+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # ChangeLog for MapMaker
 
+## MapMaker 0.3
+
+## 0.3.0 (26 Apr 2014)
+
+* new binary: akagoria-map
+  special binary to generate a map suitable for the game Akagoria
+
 ## MapMaker 0.2
 
 ## 0.2.0 (02 Mar 2014)
@@ -19,5 +26,5 @@
 
 ### 0.1.0 (22 Feb 2014)
 
-* Initial version
+* initial version